Extract Fasterer::Config from Fasterer::FileTraverser
